Complete correlation ID propagation for all token services
- Add IHttpContextAccessor to ARC200, ARC3, ASA, ARC1400 token services
- Populate CorrelationId in all token audit log entries
- Fix test mocks to include HttpContextAccessor parameter
- Enable end-to-end request tracing across all token standards

+++ BiatecTokensApi/Services/ASATokenService.cs
@@ -493,7 +493,7 @@ namespace BiatecTokensApi.Services
                     TokenName = request.Name,
                     TokenSymbol = request.UnitName,
                     TotalSupply = request.TotalSupply.ToString(),
-                    Decimals = request.Decimals,
+                    Decimals = (int)request.Decimals,
                     DeployedBy = deployedBy,
                     DeployedAt = DateTime.UtcNow,
                     Success = success,
